refactor(constants): sync AI provider constants with @blocklet/aigne-hub and remove deprecated file
