+++ blocklets/core/api/src/libs/constants.ts
@@ -1,3 +1,5 @@
+// NOTE: Need to keep in sync with @blocklet/aigne-hub/api/constants.ts
+
 export const AI_PROVIDERS = {
   OPENAI: 'openai',
   ANTHROPIC: 'anthropic',
@@ -11,25 +13,8 @@ export const AI_PROVIDERS = {
   POE: 'poe',
 } as const;
 
-export const AI_PROVIDER_DISPLAY_NAMES: Record<string, string> = {
-  [AI_PROVIDERS.OPENAI]: 'OpenAI',
-  [AI_PROVIDERS.ANTHROPIC]: 'Anthropic',
-  [AI_PROVIDERS.BEDROCK]: 'Amazon Bedrock',
-  [AI_PROVIDERS.DEEPSEEK]: 'DeepSeek',
-  [AI_PROVIDERS.GOOGLE]: 'Google',
-  [AI_PROVIDERS.OLLAMA]: 'Ollama',
-  [AI_PROVIDERS.OPENROUTER]: 'OpenRouter',
-  [AI_PROVIDERS.XAI]: 'xAI',
-  [AI_PROVIDERS.DOUBAO]: 'Doubao',
-  [AI_PROVIDERS.POE]: 'Poe',
-};
-
 export type AIProviderType = (typeof AI_PROVIDERS)[keyof typeof AI_PROVIDERS];
 
 export const AI_PROVIDER_VALUES = Object.values(AI_PROVIDERS);
 
 export const SUPPORTED_PROVIDERS_SET = new Set(AI_PROVIDER_VALUES);
-
-export const isValidAIProvider = (provider: string): provider is AIProviderType => {
-  return SUPPORTED_PROVIDERS_SET.has(provider as AIProviderType);
-};
