+++ packages/ai-kit/src/api/index.ts
@@ -1,3 +1,4 @@
 export * from './ai-kit';
 export * from './error';
 export * from './app';
+export * from './constants';
